make sure file operations and subprocess are synchronized

+++ src/launcher.py
@@ -18,10 +18,13 @@ class Launcher:
 
     def launch(self):
         print("Launch projects")
-        print("copy base project tp running environment")
-        # does it have to be asynchronized?
-        shutil.rmtree(self.run_project_path)
-        # does it have to be asynchronized?
+        print("copy base project to running environment")
+        # make sure it is synchronized
+        if os.path.isdir(self.run_project_path):
+            shutil.rmtree(self.run_project_path)
+        while os.path.isdir(self.run_project_path):
+            pass
+        # is it synchronized?
         shutil.copytree(self.base_project_path, self.run_project_path)
 
         print("change cwd to " + self.run_project_path)
@@ -33,14 +36,25 @@ class Launcher:
             if command.split(" ")[0] == "cd":
                 base_project_dir = os.path.join(base_project_dir, command.split(" ")[1])
                 continue
-
+            # subprocess.run is synchronized
             base_run = subprocess.run(command.split(), cwd=base_project_dir, env=self.env, capture_output=True, text=True)
             # process standard output and error output
             if base_run.stdout:
                 sys.stdout.write(base_run.stdout)
             if base_run.stderr:
                 sys.stderr.write(base_run.stderr)
-            # if asynchronized, process
+
+        print("copy buggy project to running envrionment")
+        # make sure it is synchronized
+        if os.path.isdir(self.run_project_path):
+            shutil.rmtree(self.run_project_path)
+        while os.path.isdir(self.run_project_path):
+            pass
+        # is it synchronized?
+        shutil.copytree(self.buggy_project_path, self.run_project_path)
+
+        print("change cwd to " + self.run_project_path)
+        os.chdir(self.run_project_path)
 
         print("executing buggy project")
         buggy_project_dir = self.run_project_path
@@ -48,7 +62,6 @@ class Launcher:
             if command.split(" ")[0] == "cd":
                 buggy_project_dir = os.path.join(buggy_project_dir, command.split(" ")[1])
                 continue
-
             buggy_run = subprocess.run(command.split(), cwd=buggy_project_dir, env=self.env, capture_output=True, text=True)
             # process standard output and error output
             if buggy_run.stdout:
@@ -57,6 +70,9 @@ class Launcher:
                 sys.stderr.write(buggy_run.stderr)
 
         os.chdir(self.cwd)
-        # does it have to be asynchronized?
-        shutil.rmtree(self.run_project_path)
-        print("finish peoject")
+        if os.path.isdir(self.run_project_path):
+            shutil.rmtree(self.run_project_path)
+        while os.path.isdir(self.run_project_path):
+            pass
+
+        print("finish project")
